move functions to separate file
